Add apartments model

+++ migrations/2023-07-06-200041_create_apartments_table/up.sql
@@ -3,7 +3,7 @@ CREATE TABLE apartments (
     location_id INT,
     location_level INT,
     location_name TEXT,
-    size INT,
+    size FLOAT,
     rooms INT,
     price TEXT,
     additional_costs INT
